Add dirty cache

+++ verifier/main.py
@@ -29,25 +29,27 @@ async def get_original_video(url):
         return await client.get(url)
 
 
+cache = {}
+
 @app.get('/verify/{cid}/{direction}/{second_direction}')
 async def verify(cid: str, direction: int, second_direction: int):
-    resp = await get_original_video(base_url.format(cid))
-    original_file_path = f'/tmp/{cid}.mp4'
-    async with aiofiles.open(original_file_path, 'wb') as out_file:
-        await out_file.write(resp.content)
-    loop = asyncio.get_running_loop()
-    is_verified, direction_time, second_direction_time, rotateCode = verify_video(original_file_path, direction, second_direction)
-    await aiofiles.os.remove(original_file_path)
-
-    return JSONResponse(
-        status_code=200,
-        content={
+    content = cache.get(cid)
+    if not content:
+        resp = await get_original_video(base_url.format(cid))
+        original_file_path = f'/tmp/{cid}.mp4'
+        async with aiofiles.open(original_file_path, 'wb') as out_file:
+            await out_file.write(resp.content)
+        loop = asyncio.get_running_loop()
+        is_verified, direction_time, second_direction_time, rotateCode = verify_video(original_file_path, direction, second_direction)
+        await aiofiles.os.remove(original_file_path)
+        content = {
             'is_verified': is_verified,
             'direction_time': direction_time,
             'second_direction_time': second_direction_time,
             'rotate_code': rotateCode,
-        },
-    )
+        }
+
+    return JSONResponse(status_code=200, content=content)
 
 
 if __name__ == "__main__":
